Update according to @sverhoeven's review. Split up module.
